fix data type

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -9,7 +9,7 @@
  */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int diff_bits;
+	unsigned long int diff_bits;
 	int count;
 
 	diff_bits = n ^ m;
